Narona_ASI/pi-zero-camera/requirements.txt
flask>=3.0.0
picamera2>=0.3.12
Pillow>=10.0.0
